Move wine property from game to preferences
It doesn't make sense to store a separate path for Wine for every game, so I moved it into the global SourceOps preferences.
