+++ addon/props/game_props.py
@@ -23,13 +23,6 @@ class SOURCEOPS_GameProps(bpy.types.PropertyGroup):
         update=utils.game.update_bin,
     )
 
-    wine: bpy.props.StringProperty(
-        name='Wine',
-        description='Path to your wine installation',
-        subtype='FILE_PATH',
-        update=utils.game.update_wine,
-    )
-
     modelsrc: bpy.props.StringProperty(
         name='ModelSrc',
         description='Path to your modelsrc folder',
