+++ addon/ui/panels.py
@@ -50,8 +50,6 @@ class SOURCEOPS_PT_MainPanel(bpy.types.Panel):
                 col.prop(game, 'name')
                 col.prop(game, 'game')
                 col.prop(game, 'bin')
-                if os.name == 'posix':
-                    col.prop(game, 'wine')
                 col.prop(game, 'modelsrc')
                 col.prop(game, 'models')
                 col.prop(game, 'mapsrc')
